Add controller upgrade management
Memoized functions in the Goals object were unnecessary since the cached
target arguments are invalidated on the next tick, causing a memory
leak. The only Goal type where this isn't the case is in NullGoal since
it has no constructor arguments to memoize.

The findStorages method of MineOfficial had to be factored out into
Official so that ControllerOfficial could use it. That method then had
to be generalized with an argument to favor either empty or full
storages.